Created Interface for Inventory
