+++ Shopping_Cart/Inventory/IInventory.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ShoppingCart
+{
+    public interface IInventory
+    {
+        public void DisplayMenu();
+    }
+}
